# outputs/security/dynamic/raw/openapi.json
{"components": {"schemas": {"AccessLevel": {"enum": ["metadata_only", "aggregate_analysis", "controlled_export"], "title": "AccessLevel", "type": "string"}, "AccessRequestCreate": {"additionalProperties": false, "properties": {"dataset_id": {"maxLength": 120, "minLength": 1, "pattern": "^[A-Za-z0-9_.:-]+$", "title": "Dataset Id", "type": "string"}, "requested_access_level": {"$ref": "#/components/schemas/AccessLevel"}, "research_purpose": {"maxLength": 500, "minLength": 1, "title": "Research Purpose", "type": "string"}}, "required": ["dataset_id", "research_purpose", "requested_access_level"], "title": "AccessRequestCreate", "type": "object"}, "AccessRequestResponse": {"additionalProperties": false, "properties": {"dataset_id": {"title": "Dataset Id", "type": "string"}, "decision_reason": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Decision Reason"}, "request_id": {"title": "Request Id", "type": "string"}, "requested_access_level": {"$ref": "#/components/schemas/AccessLevel"}, "requester_id": {"title": "Requester Id", "type": "string"}, "research_purpose": {"title": "Research Purpose", "type": "string"}, "reviewed_at": {"anyOf": [{"format": "date-time", "type": "string"}, {"type": "null"}], "title": "Reviewed At"}, "reviewed_by": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Reviewed By"}, "status": {"$ref": "#/components/schemas/AccessRequestStatus"}, "submitted_at": {"format": "date-time", "title": "Submitted At", "type": "string"}}, "required": ["request_id", "dataset_id", "requester_id", "research_purpose", "requested_access_level", "status", "submitted_at", "reviewed_at", "reviewed_by", "decision_reason"], "title": "AccessRequestResponse", "type": "object"}, "AccessRequestStatus": {"enum": ["pending", "approved", "rejected", "withdrawn"], "title": "AccessRequestStatus", "type": "string"}, "ActorRole": {"enum": ["researcher", "approver", "data_custodian", "security_auditor", "application_admin"], "title": "ActorRole", "type": "string"}, "ApprovalDecisionRequest": {"additionalProperties": false, "properties": {"decision_reason": {"description": "Approval reason recorded by the reviewer.", "maxLength": 500, "minLength": 1, "title": "Decision Reason", "type": "string"}}, "required": ["decision_reason"], "title": "ApprovalDecisionRequest", "type": "object"}, "AuditEventResponse": {"additionalProperties": false, "properties": {"actor_id": {"title": "Actor Id", "type": "string"}, "actor_role": {"$ref": "#/components/schemas/ActorRole"}, "correlation_id": {"title": "Correlation Id", "type": "string"}, "details": {"additionalProperties": {"type": "string"}, "title": "Details", "type": "object"}, "event_id": {"title": "Event Id", "type": "string"}, "event_type": {"$ref": "#/components/schemas/AuditEventType"}, "outcome": {"$ref": "#/components/schemas/AuditOutcome"}, "resource_id": {"title": "Resource Id", "type": "string"}, "resource_type": {"title": "Resource Type", "type": "string"}, "timestamp": {"format": "date-time", "title": "Timestamp", "type": "string"}}, "required": ["event_id", "event_type", "actor_id", "actor_role", "resource_type", "resource_id", "outcome", "timestamp", "correlation_id", "details"], "title": "AuditEventResponse", "type": "object"}, "AuditEventType": {"enum": ["authentication_succeeded", "authentication_failed", "authorisation_denied", "dataset_viewed", "access_request_submitted", "access_request_viewed", "access_request_approved", "access_request_rejected", "self_approval_denied", "invalid_workflow_transition_attempted", "audit_events_viewed", "rate_limit_exceeded"], "title": "AuditEventType", "type": "string"}, "AuditOutcome": {"enum": ["success", "failure"], "title": "AuditOutcome", "type": "string"}, "DatasetResponse": {"additionalProperties": false, "properties": {"access_level": {"$ref": "#/components/schemas/AccessLevel"}, "created_at": {"format": "date-time", "title": "Created At", "type": "string"}, "dataset_id": {"title": "Dataset Id", "type": "string"}, "description": {"title": "Description", "type": "string"}, "name": {"title": "Name", "type": "string"}, "research_domain": {"title": "Research Domain", "type": "string"}, "sensitivity_classification": {"$ref": "#/components/schemas/SensitivityClassification"}, "status": {"$ref": "#/components/schemas/DatasetStatus"}, "updated_at": {"format": "date-time", "title": "Updated At", "type": "string"}}, "required": ["dataset_id", "name", "description", "research_domain", "sensitivity_classification", "access_level", "status", "created_at", "updated_at"], "title": "DatasetResponse", "type": "object"}, "DatasetStatus": {"enum": ["active", "retired"], "title": "DatasetStatus", "type": "string"}, "ErrorBody": {"additionalProperties": false, "properties": {"code": {"title": "Code", "type": "string"}, "correlation_id": {"title": "Correlation Id", "type": "string"}, "message": {"title": "Message", "type": "string"}}, "required": ["code", "message", "correlation_id"], "title": "ErrorBody", "type": "object"}, "ErrorResponse": {"additionalProperties": false, "properties": {"error": {"$ref": "#/components/schemas/ErrorBody"}}, "required": ["error"], "title": "ErrorResponse", "type": "object"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "title": "Detail", "type": "array"}}, "title": "HTTPValidationError", "type": "object"}, "HealthResponse": {"additionalProperties": false, "properties": {"service": {"title": "Service", "type": "string"}, "status": {"title": "Status", "type": "string"}, "version": {"title": "Version", "type": "string"}}, "required": ["status", "service", "version"], "title": "HealthResponse", "type": "object"}, "RejectionDecisionRequest": {"additionalProperties": false, "properties": {"decision_reason": {"description": "Rejection reason recorded by the reviewer.", "maxLength": 500, "minLength": 1, "title": "Decision Reason", "type": "string"}}, "required": ["decision_reason"], "title": "RejectionDecisionRequest", "type": "object"}, "SensitivityClassification": {"enum": ["synthetic_controlled", "synthetic_restricted"], "title": "SensitivityClassification", "type": "string"}, "ValidationError": {"properties": {"ctx": {"title": "Context", "type": "object"}, "input": {"title": "Input"}, "loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "title": "Location", "type": "array"}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}, "required": ["loc", "msg", "type"], "title": "ValidationError", "type": "object"}}, "securitySchemes": {"HTTPBearer": {"scheme": "bearer", "type": "http"}}}, "info": {"description": "Synthetic, non-identifiable portfolio API for controlled genomic research dataset access workflows.", "title": "Genomic Research Access API", "version": "0.1.0"}, "openapi": "3.1.0", "paths": {"/api/v1/access-requests": {"get": {"operationId": "list_access_requests_api_v1_access_requests_get", "responses": {"200": {"content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/AccessRequestResponse"}, "title": "Response List Access Requests Api V1 Access Requests Get", "type": "array"}}}, "description": "Successful Response"}, "401": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}, "description": "Authentication is missing, expired or invalid."}, "403": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}, "description": "The authenticated principal is not authorised for this action."}, "429": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}, "description": "The request rate limit was exceeded."}}, "security": [{"HTTPBearer": []}], "summary": "List Access Requests", "tags": ["access requests"]}, "post": {"operationId": "create_access_request_api_v1_access_requests_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccessRequestCreate"}}}, "required": true}, "responses": {"201": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccessRequestResponse"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "security": [{"HTTPBearer": []}], "summary": "Create Access Request", "tags": ["access requests"]}}, "/api/v1/access-requests/{request_id}": {"get": {"operationId": "get_access_request_api_v1_access_requests__request_id__get", "parameters": [{"example": "schemathesis-access-request-001", "in": "path", "name": "request_id", "required": true, "schema": {"examples": ["schemathesis-access-request-001"], "maxLength": 120, "minLength": 1, "pattern": "^[A-Za-z0-9_.:-]+$", "title": "Request Id", "type": "string"}}], "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccessRequestResponse"}}}, "description": "Successful Response"}, "401": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}, "description": "Authentication is missing, expired or invalid."}, "403": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}, "description": "The authenticated principal is not authorised for this action."}, "404": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}, "description": "The requested resource was not found or is not visible to the principal."}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}, "429": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}, "description": "The request rate limit was exceeded."}}, "security": [{"HTTPBearer": []}], "summary": "Get Access Request", "tags": ["access requests"]}}, "/api/v1/access-requests/{request_id}/approve": {"post": {"operationId": "approve_access_request_api_v1_access_requests__request_id__approve_post", "parameters": [{"in": "path", "name": "request_id", "required": true, "schema": {"title": "Request Id", "type": "string"}}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ApprovalDecisionRequest"}}}, "required": true}, "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccessRequestResponse"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "security": [{"HTTPBearer": []}], "summary": "Approve Access Request", "tags": ["access requests"]}}, "/api/v1/access-requests/{request_id}/reject": {"post": {"operationId": "reject_access_request_api_v1_access_requests__request_id__reject_post", "parameters": [{"in": "path", "name": "request_id", "required": true, "schema": {"title": "Request Id", "type": "string"}}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RejectionDecisionRequest"}}}, "required": true}, "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccessRequestResponse"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "security": [{"HTTPBearer": []}], "summary": "Reject Access Request", "tags": ["access requests"]}}, "/api/v1/audit-events": {"get": {"description": "Return local in-memory audit events for Milestone 1 demonstration only.", "operationId": "list_audit_events_api_v1_audit_events_get", "responses": {"200": {"content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/AuditEventResponse"}, "title": "Response List Audit Events Api V1 Audit Events Get", "type": "array"}}}, "description": "Successful Response"}, "401": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}, "description": "Authentication is missing, expired or invalid."}, "403": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}, "description": "The authenticated principal is not authorised for this action."}, "429": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}, "description": "The request rate limit was exceeded."}}, "security": [{"HTTPBearer": []}], "summary": "List Audit Events", "tags": ["audit events"]}}, "/api/v1/datasets": {"get": {"operationId": "list_datasets_api_v1_datasets_get", "responses": {"200": {"content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/DatasetResponse"}, "title": "Response List Datasets Api V1 Datasets Get", "type": "array"}}}, "description": "Successful Response"}}, "security": [{"HTTPBearer": []}], "summary": "List Datasets", "tags": ["datasets"]}}, "/api/v1/datasets/{dataset_id}": {"get": {"operationId": "get_dataset_api_v1_datasets__dataset_id__get", "parameters": [{"in": "path", "name": "dataset_id", "required": true, "schema": {"title": "Dataset Id", "type": "string"}}], "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/DatasetResponse"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "security": [{"HTTPBearer": []}], "summary": "Get Dataset", "tags": ["datasets"]}}, "/health": {"get": {"operationId": "health_health_get", "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HealthResponse"}}}, "description": "Successful Response"}}, "summary": "Health", "tags": ["health"]}}}}